+++ src/main/scala/net/psforever/objects/ce/DeployableBehavior.scala
@@ -2,10 +2,9 @@
 package net.psforever.objects.ce
 
 import akka.actor.{Actor, ActorRef, Cancellable}
-import net.psforever.objects.guid.GUIDTask
+import net.psforever.objects.guid.{GUIDTask, TaskWorkflow}
 import net.psforever.objects._
 import net.psforever.objects.definition.DeployAnimation
-import net.psforever.objects.guid.actor.TaskWorkflow
 import net.psforever.objects.zones.Zone
 import net.psforever.packet.game._
 import net.psforever.services.Service
