+++ src/main/scala/net/psforever/objects/guid/source/Key.scala
@@ -2,8 +2,7 @@
 package net.psforever.objects.guid.source
 
 import net.psforever.objects.entity.IdentifiableEntity
-import net.psforever.objects.guid.AvailabilityPolicy
-import net.psforever.objects.guid.key.Monitor
+import net.psforever.objects.guid.key.{AvailabilityPolicy, Monitor}
 
 private class Key extends Monitor {
   var policy: AvailabilityPolicy = AvailabilityPolicy.Available
