+++ src/main/scala/net/psforever/objects/guid/uns/UnregisteredEntity.scala
@@ -1,5 +1,5 @@
 // Copyright (c) 2021 PSForever
-package net.psforever.objects.guid.actor
+package net.psforever.objects.guid.uns
 
 import net.psforever.objects.entity.IdentifiableEntity
 import net.psforever.objects.guid.NumberPoolHub
@@ -10,3 +10,5 @@ final case class UnregisteredEntity(
                                      guid_system: NumberPoolHub,
                                      number: Int
                                    )
+
+final case class AlreadyUnregisteredEntity(msg: UnregisteredEntity)
