+++ src/main/scala/net/psforever/objects/serverobject/pad/VehicleSpawnControl.scala
@@ -4,8 +4,7 @@ package net.psforever.objects.serverobject.pad
 import akka.actor.{Cancellable, Props}
 import net.psforever.objects.avatar.SpecialCarry
 import net.psforever.objects.entity.WorldEntity
-import net.psforever.objects.guid.GUIDTask
-import net.psforever.objects.guid.actor.TaskWorkflow
+import net.psforever.objects.guid.{GUIDTask, TaskWorkflow}
 import net.psforever.objects.serverobject.affinity.{FactionAffinity, FactionAffinityBehavior}
 import net.psforever.objects.serverobject.pad.process.{VehicleSpawnControlBase, VehicleSpawnControlConcealPlayer}
 import net.psforever.objects.zones.{Zone, ZoneAware, Zoning}
