+++ src/main/scala/net/psforever/services/vehicle/support/TurretUpgrader.scala
@@ -3,9 +3,8 @@ package net.psforever.services.vehicle.support
 
 import akka.actor.Cancellable
 import net.psforever.objects.equipment.EquipmentSlot
-import net.psforever.objects.guid.actor.{Task, TaskBundle, TaskWorkflow}
 import net.psforever.objects.{AmmoBox, Default, PlanetSideGameObject, Tool}
-import net.psforever.objects.guid.GUIDTask
+import net.psforever.objects.guid.{GUIDTask, Task, TaskBundle, TaskWorkflow}
 import net.psforever.objects.serverobject.PlanetSideServerObject
 import net.psforever.objects.serverobject.turret.{FacilityTurret, TurretUpgrade, WeaponTurret}
 import net.psforever.objects.vehicles.MountedWeapons
