+++ src/test/scala/objects/guidtask/GUIDTaskTest.scala
@@ -4,14 +4,13 @@ package objects.guidtask
 import java.util.logging.LogManager
 
 import scala.util.Success
-import akka.actor.{ActorContext, ActorRef, ActorSystem, Props}
+import akka.actor.{ActorRef, ActorSystem, Props}
 import akka.testkit.TestProbe
 import net.psforever.objects.entity.IdentifiableEntity
-import net.psforever.objects.guid.actor.{NumberPoolActor, UniqueNumberSystem}
 import net.psforever.objects.guid.selector.RandomSelector
 import net.psforever.objects.guid.source.MaxNumberSource
-import net.psforever.objects.guid.NumberPoolHub
-import net.psforever.objects.guid.actor.Task
+import net.psforever.objects.guid.uns.NumberPoolActor
+import net.psforever.objects.guid.{NumberPoolHub, Task, UniqueNumberOps}
 
 import scala.concurrent.Future
 
@@ -28,12 +27,10 @@ object GUIDTaskTest {
     override def isSuccessful() : Boolean = false
   }
 
-  def CommonTestSetup(implicit system: ActorSystem): (NumberPoolHub, ActorRef, ActorRef, TestProbe) = {
-    import akka.actor.Props
-    import akka.routing.RandomPool
+  def CommonTestSetup(implicit system: ActorSystem): (NumberPoolHub, UniqueNumberOps, ActorRef, TestProbe) = {
     import akka.testkit.TestProbe
 
-    val guid: NumberPoolHub = new NumberPoolHub(new MaxNumberSource(110))
+    val guid: NumberPoolHub = new NumberPoolHub(new MaxNumberSource(90))
     guid.AddPool("players", (1 to 10).toList).Selector = new RandomSelector
     guid.AddPool("lockers", (11 to 20).toList).Selector = new RandomSelector
     guid.AddPool("ammo", (21 to 30).toList).Selector = new RandomSelector
@@ -42,18 +39,13 @@ object GUIDTaskTest {
     guid.AddPool("terminals", (51 to 60).toList).Selector = new RandomSelector
     guid.AddPool("items", (61 to 70).toList).Selector = new RandomSelector
     guid.AddPool("deployables", (71 to 80).toList).Selector = new RandomSelector
-    val func: (ActorContext, NumberPoolHub) => Map[String, ActorRef] =
-      UniqueNumberSystem.AllocateNumberPoolActors
-    val uns = system.actorOf(
-      RandomPool(25).props(Props(classOf[UniqueNumberSystem], guid, func)),
-      "uns"
-    )
+    val uns = new UniqueNumberOps(guid, AllocateNumberPoolActors(guid)(system))
     LogManager.getLogManager.reset() //suppresses any internal loggers created by the above elements
     (guid, uns, ActorRef.noSender, TestProbe())
   }
 
   /**
-    * @see `UniqueNumberSystem.AllocateNumberPoolActors(NumberPoolHub)(implicit ActorContext)`
+    * @see `UniqueNumberSetup.AllocateNumberPoolActors(NumberPoolHub)(implicit ActorContext)`
     */
   def AllocateNumberPoolActors(poolSource: NumberPoolHub)(implicit system: ActorSystem): Map[String, ActorRef] = {
     poolSource.Pools
